Renamed the 'implementation' package to 'internal' to follow Android conventions. Changed the visibility of some getModule() methods that can be package-private instead of public. Moved an exception class to its own package for the sake of consistency. Removed the AdditionalProperties and Flattening serializers and deserializers since they are not required at the moment.
