--- sdk/core/azure-core/src/main/java/com/azure/android/core/implementation/util/serializer/AdditionalPropertiesDeserializer.java
@@ -1,152 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-package com.azure.android.core.implementation.util.serializer;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.core.JsonFactory;
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.databind.BeanDescription;
-import com.fasterxml.jackson.databind.DeserializationConfig;
-import com.fasterxml.jackson.databind.DeserializationContext;
-import com.fasterxml.jackson.databind.JsonDeserializer;
-import com.fasterxml.jackson.databind.JsonMappingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.deser.BeanDeserializerModifier;
-import com.fasterxml.jackson.databind.deser.ResolvableDeserializer;
-import com.fasterxml.jackson.databind.deser.std.StdDeserializer;
-import com.fasterxml.jackson.databind.module.SimpleModule;
-import com.fasterxml.jackson.databind.node.ObjectNode;
-
-import java.io.IOException;
-import java.lang.reflect.Field;
-import java.util.ArrayList;
-import java.util.List;
-
-/**
- * Custom serializer for deserializing complex types with additional properties. If a complex type has a property
- * named "additionalProperties" with serialized name empty ("") of type Map&lt;String, Object&gt;, all extra
- * properties on the payload will be stored in this map.
- */
-final class AdditionalPropertiesDeserializer extends StdDeserializer<Object> implements ResolvableDeserializer {
-    private static final long serialVersionUID = 700052863615540646L;
-
-    /**
-     * The default mapperAdapter for the current type.
-     */
-    private final JsonDeserializer<?> defaultDeserializer;
-
-    /**
-     * The object mapper for default deserializations.
-     */
-    private final ObjectMapper mapper;
-
-    /**
-     * Creates an instance of {@link FlatteningDeserializer}.
-     *
-     * @param vc                  Handled type.
-     * @param defaultDeserializer The default JSON mapperAdapter.
-     * @param mapper              The object mapper for default deserializations.
-     */
-    protected AdditionalPropertiesDeserializer(Class<?> vc, JsonDeserializer<?> defaultDeserializer,
-                                               ObjectMapper mapper) {
-        super(vc);
-
-        this.defaultDeserializer = defaultDeserializer;
-        this.mapper = mapper;
-    }
-
-    /**
-     * Gets a module wrapping this serializer as an adapter for the Jackson {@link ObjectMapper}.
-     *
-     * @param mapper The object mapper for default deserializations.
-     * @return A simple module to be plugged onto Jackson {@link ObjectMapper}.
-     */
-    public static SimpleModule getModule(final ObjectMapper mapper) {
-        SimpleModule module = new SimpleModule();
-
-        module.setDeserializerModifier(new BeanDeserializerModifier() {
-            @Override
-            public JsonDeserializer<?> modifyDeserializer(DeserializationConfig config,
-                                                          BeanDescription beanDesc,
-                                                          JsonDeserializer<?> deserializer) {
-                for (Class<?> c : getAllClasses(beanDesc.getBeanClass())) {
-                    Field[] fields = c.getDeclaredFields();
-
-                    for (Field field : fields) {
-                        if ("additionalProperties".equalsIgnoreCase(field.getName())) {
-                            JsonProperty property = field.getAnnotation(JsonProperty.class);
-
-                            if (property != null && property.value().isEmpty()) {
-                                return new AdditionalPropertiesDeserializer(beanDesc.getBeanClass(), deserializer,
-                                    mapper);
-                            }
-                        }
-                    }
-                }
-
-                return deserializer;
-            }
-        });
-
-        return module;
-    }
-
-    @Override
-    public Object deserialize(JsonParser jsonParser, DeserializationContext context) throws IOException {
-        ObjectNode root = mapper.readTree(jsonParser);
-        ObjectNode copy = root.deepCopy();
-
-        // Compare top level fields and keep only missing fields
-        final Class<?> tClass = this.defaultDeserializer.handledType();
-
-        for (Class<?> c : getAllClasses(tClass)) {
-            Field[] fields = c.getDeclaredFields();
-
-            for (Field field : fields) {
-                // JaCoCo adds synthetic fields for instrumentation.
-                // It's recommended to skip fields that are marked synthetic.
-                // https://www.eclemma.org/jacoco/trunk/doc/faq.html
-                // https://github.com/jacoco/jacoco/issues/168
-                if (field.isSynthetic()) {
-                    continue;
-                }
-
-                JsonProperty property = field.getAnnotation(JsonProperty.class);
-                String key = property.value().split("((?<!\\\\))\\.")[0];
-
-                if (!key.isEmpty()) {
-                    if (copy.has(key)) {
-                        copy.remove(key);
-                    }
-                }
-            }
-        }
-
-        // Put into additional properties
-        root.put("additionalProperties", copy);
-
-        JsonParser parser = new JsonFactory().createParser(root.toString());
-
-        parser.nextToken();
-
-        return defaultDeserializer.deserialize(parser, context);
-    }
-
-    @Override
-    public void resolve(DeserializationContext context) throws JsonMappingException {
-        ((ResolvableDeserializer) defaultDeserializer).resolve(context);
-    }
-
-    private static List<Class<?>> getAllClasses(Class<?> clazz) {
-        List<Class<?>> types = new ArrayList<>();
-
-        while (clazz != null) {
-            types.add(clazz);
-
-            clazz = clazz.getSuperclass();
-        }
-
-        return types;
-    }
-}
