+++ sdk/core/azure-core/src/main/java/com/azure/android/core/internal/util/ExceptionUtils.java
@@ -1,11 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.android.core.implementation.util;
+package com.azure.android.core.internal.util;
 
-import com.azure.android.core.exception.HttpResponseException;
-import com.azure.android.core.implementation.util.serializer.SerializerAdapter;
-import com.azure.android.core.implementation.util.serializer.SerializerEncoding;
+import com.azure.android.core.http.exception.HttpResponseException;
+import com.azure.android.core.internal.util.serializer.SerializerAdapter;
+import com.azure.android.core.internal.util.serializer.SerializerEncoding;
 
 import java.io.IOException;
 import java.lang.reflect.Constructor;
