+++ sdk/core/azure-core/src/main/java/com/azure/android/core/internal/util/serializer/SerializerEncoding.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.android.core.implementation.util.serializer;
+package com.azure.android.core.internal.util.serializer;
 
 import okhttp3.Headers;
